requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
seaborn
joblib
Pillow